requirements.txt
Django==4.2.7
pywebview==4.4
Pillow==10.1.0
python-decouple==3.8
pyinstaller==6.1.0
pypdf>=4.0.0
python-docx>=1.1.0
reportlab>=4.0.0
djangorestframework>=3.14.0
django-apscheduler>=0.6.2
